Extract methods to individual classes
