+++ lib/mlb_rb/errors.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module MlbRb
+  class DateError < StandardError; end
+end
