+++ spec/mlb_rb_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "./lib/mlb_rb/games"
+
 RSpec.describe MlbRb do
   it "has a version number" do
     expect(MlbRb::VERSION).not_to be nil
@@ -10,54 +12,6 @@ RSpec.describe MlbRb do
       healthcheck = MlbRb.healthcheck
 
       expect(healthcheck).to eq("Yup, everything is fine")
-    end
-  end
-
-  describe ".games_for_date" do
-    let(:file) { File.read("./spec/fixtures/api_v1_schedule_2022-03-23.json") }
-    it "gets the games for a specific date" do
-      expect(Net::HTTP).to receive(:get)
-        .with("statsapi.mlb.com", "/api/v1/schedule/games/?sportId=1&date=03/23/2022")
-        .and_return(file)
-
-      games = MlbRb.games_for_date({date: {year: 2022, month: 3, day: 23}})
-
-      expect(games.length).to eq(14)
-
-      first_game = games.first
-      expect(first_game.game_pk).to eq(706964)
-      expect(first_game.home_team.name).to eq("Detroit Tigers")
-      expect(first_game.away_team.name).to eq("Pittsburgh Pirates")
-      expect(first_game.home_score).to eq(6)
-      expect(first_game.away_score).to eq(6)
-      expect(first_game.game_date).to eq(Date.new(2022, 0o3, 23))
-    end
-  end
-
-  describe ".games_for_date_range" do
-    let(:file) { File.read("./spec/fixtures/api_v1_schedule_2022-03-23_2022-03-24.json") }
-    it "gets the games for a specific date range" do
-      expect(Net::HTTP).to receive(:get)
-        .with("statsapi.mlb.com", "/api/v1/schedule/games/?sportId=1&startDate=03/23/2022&endDate=03/24/2022")
-        .and_return(file)
-
-      games = MlbRb.games_for_date_range(
-        {
-          start_date: {year: 2022, month: 3, day: 23},
-          end_date: {year: 2022, month: 3, day: 24}
-        }
-      )
-
-      expect(games.length).to eq(28)
-
-      last_game = games.last
-      expect(last_game.game_pk).to eq(706798)
-      expect(last_game.home_team.name).to eq("Toronto Blue Jays")
-      expect(last_game.away_team.name).to eq("Atlanta Braves")
-
-      expect(last_game.home_score).to eq(nil)
-      expect(last_game.away_score).to eq(nil)
-      expect(last_game.game_date).to eq(Date.new(2022, 0o3, 24))
     end
   end
 end
